feat: add settings module
