+++ mobile/lib/core/router/app_router.dart
@@ -12,7 +12,7 @@ import '../../features/notes/presentation/trash_screen.dart';
 import '../../features/notes/presentation/archive_screen.dart';
 import '../presentation/splash_screen.dart';
 import '../presentation/server_config_screen.dart';
-import '../presentation/settings_screen.dart';
+import '../../features/settings/presentation/settings_screen.dart';
 import '../../features/auth/presentation/auth_controller.dart';
 import '../network/server_config_provider.dart';
 import 'app_routes.dart';
